refactor: arquitecutra screaming
