--- frontend/src/components/First.tsx
@@ -1,5 +1,0 @@
-function First() {
-  return <div></div>;
-}
-
-export default First;
