--- frontend/src/constants/theme.ts
@@ -1,25 +1,0 @@
-export const theme = {
-  colors: {
-    background: "#0f172a", // Fondo principal del formulario
-    surface: "#1e293b", // Fondo secundario del contenedor
-    gradientStart: "#2757a5", // Azul degradado botón
-    gradientEnd: "#6952bb", // Violeta degradado botón
-    border: "#4d6485",
-
-    primary: "#3b82f6", // Azul primario
-    secondary: "#8b5cf6", // Violeta secundario
-    accent: "#facc15", // Amarillo para elementos destacados
-
-    text: "#f1f5f9", // Texto principal
-    placeholder: "#94a3b8", // Placeholder
-    muted: "#64748b", // Texto menos importante
-
-    inputBg: "#334155", // Fondo inputs
-    inputBorder: "#475569", // Borde inputs
-    error: "#ef4444", // Errores / validaciones
-
-    overlay: "rgba(0, 0, 0, 0.6)", // Fondo oscuro semitransparente
-  },
-
-  borderRadius: "0.75rem",
-};
