+++ frontend/src/main.tsx
@@ -1,6 +1,5 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './styles/global.css'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
