+++ frontend/src/modules/Dashboard/components/Totals.tsx
@@ -1,0 +1,56 @@
+import { Box, type SxProps, type Theme } from '@mui/material';
+import { TopCategoriesCard } from './TopCategoriesCard/TopCategoriesCard';
+import { TotalsBalanceCard } from './TotalsBalanceCard'; // Asegúrate de importar tu card de totales
+
+interface Expense {
+  id: number;
+  amount: number;
+  date: string;
+  category: string;
+  payment_method: string;
+  description: string;
+  user_id: number;
+}
+
+interface TotalsProps {
+  totalGastos: number;
+  totalIngresos: number;
+  balance: number;
+  expenses: Expense[];
+}
+
+const MAX_CATEGORIES = 5;
+const TOP_CATEGORIES_TITLE = "Top Categorías de Gasto";
+
+const totalsContainerSx: SxProps<Theme> = {
+  display: 'flex',
+  flexWrap: 'wrap',
+  gap: 2,
+  width: '100%',
+};
+
+const totalsItemSx: SxProps<Theme> = {
+  flex: { xs: '35%', md: '20%' },
+};
+
+export const Totals = ({ totalGastos, totalIngresos, balance, expenses }: TotalsProps) => {
+  return (
+    <Box sx={totalsContainerSx}>
+      <Box sx={totalsItemSx}>
+        <TotalsBalanceCard 
+          totalGastos={totalGastos}
+          totalIngresos={totalIngresos}
+          balance={balance}
+        />
+      </Box>
+
+      <Box sx={totalsItemSx}>
+        <TopCategoriesCard 
+          title={TOP_CATEGORIES_TITLE}
+          expenses={expenses}
+          maxCategories={MAX_CATEGORIES}
+        />
+      </Box>
+    </Box>
+  );
+};
