+++ frontend/src/modules/Dashboard/types/balance.ts
@@ -1,0 +1,6 @@
+
+export interface TotalsBalanceCardProps {
+  totalGastos: number;
+  totalIngresos: number;
+  balance: number;
+}
