+++ frontend/src/modules/Dashboard/types/dashed.ts
@@ -1,0 +1,13 @@
+export interface DataItem {
+  id: number;
+  amount: number;
+  date: string;
+  category: string;
+  description: string;
+  user_id: number;
+}
+
+export interface DashedLineChartProps {
+  expenses: DataItem[];
+  incomes: DataItem[];
+}
