+++ frontend/src/modules/Dashboard/types/expense.ts
@@ -1,0 +1,9 @@
+export interface Expense {
+  id: number;
+  amount: number;
+  date: string;
+  category: string;
+  payment_method: string;
+  description: string;
+  user_id: number;
+}
