+++ frontend/src/modules/Dashboard/types/pichart.ts
@@ -1,0 +1,8 @@
+import type { Expense } from "./expense";
+
+export interface PieChartProps {
+  expenses: Expense[];
+  backgroundD?: string;
+  colo?: string;
+  display?: object;
+}
