--- frontend/src/pages/Gastos.tsx
@@ -1,12 +1,0 @@
-
-
-const Gastos = () => {
-  return (
-    <div>
-      <h2>Seccion de gastos</h2>
-   
-    </div>
-  );
-};
-
-export default Gastos;
