--- frontend/src/styles/global.css
@@ -1,34 +1,0 @@
-:root {
---color-background: #0f172a;        /* Fondo principal del formulario */
---color-surface: #1e293b;           /* Fondo secundario del contenedor */
---color-gradient-start: #3b82f6;    /* Azul degradado botón */
---color-gradient-end: #8b5cf6;      /* Violeta degradado botón */
-
---color-border: #4d6485;
-
---color-primary: #2757a5;           /* Azul primario (botones, íconos activos) */
---color-secondary: #6952bb;         /* Violeta secundario */
-
---color-accent: #facc15;            /* Amarillo para elementos destacados si lo necesitás */
-
---color-text: #f1f5f9;              /* Texto principal (blanco grisáceo) */
---color-placeholder: #94a3b8;       /* Texto placeholder en inputs */
---color-muted: #64748b;             /* Texto de ayuda o menos importante */
-
---color-input-bg: #334155;          /* Fondo de los campos de texto */
---color-input-border: #475569;      /* Borde de inputs */
---color-error: #ef4444;             /* Para validaciones */
-
---color-overlay: rgba(0, 0, 0, 0.6); /* Fondos oscuros semitransparentes */
---border-radius: 0.75rem;
-}
-
-html, body {
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  background: var(--color-background);
-  color: var(--color-text);
-}
